Only replace solc-js copies in directories present during externalTests

+++ test/externalTests.sh
@@ -59,10 +59,13 @@ function test_truffle
       for d in node_modules node_modules/truffle/node_modules
       do
       (
-        cd $d
-        rm -rf solc
-        git clone --depth 1 -b v0.5.0 https://github.com/ethereum/solc-js.git solc
-        cp "$SOLJSON" solc/
+        if [ -d "$d" ]
+        then
+          cd $d
+          rm -rf solc
+          git clone --depth 1 -b v0.5.0 https://github.com/ethereum/solc-js.git solc
+          cp "$SOLJSON" solc/
+        fi
       )
       done
       if [ "$name" == "Zeppelin" -o "$name" == "Gnosis" ]; then
